backend/requirements.txt

Django==4.2.13
django-cors-headers==4.7.0
djangorestframework==3.16.0
gunicorn==22.0.0
pipenv==2024.1.0

